Release 1.0.5 1、优化文字识别速率 2、优化截图效率 3、实现root工作模式

+++ app/src/main/java/com/smart/autodaily/data/entity/CustomTypeConvert.kt
@@ -32,7 +32,7 @@ object CustomTypeConvert {
     }
     @TypeConverter
     fun toPoint(data: String): Point {
-        val (x, y) = data.split(",").map { it.toFloat() }
+        val (x, y) = data.split(",").map { it.toInt() }
         return Point(x, y)
     }
 }
